(PC-1147) Refactoring after code review suggestions
